requirements.txt
numpy 
networkx 
sentence-transformers 
faiss-cpu 
streamlit
pymupdf